refactor(api): move stage collection repositories
