+++ api/src/shared/domain/models/target-profile-management/StageCollectionUpdate.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { InvalidStageError } from '../../errors.js';
+import { InvalidStageError } from '../../../../../lib/domain/errors.js';
 
 const DEFAULT_VALUE_FIRST_SKILL = -1;
