+++ api/src/shared/domain/models/user-campaign-results/StageCollection.js
