+++ api/tests/shared/unit/domain/models/target-profile-management/StageCollectionUpdate_test.js
@@ -1,6 +1,6 @@
-import { expect, domainBuilder } from '../../../../test-helper.js';
-import { InvalidStageError } from '../../../../../lib/domain/errors.js';
-import { StageCollectionUpdate } from '../../../../../lib/domain/models/target-profile-management/StageCollectionUpdate.js';
+import { expect, domainBuilder } from '../../../../../test-helper.js';
+import { InvalidStageError } from '../../../../../../lib/domain/errors.js';
+import { StageCollectionUpdate } from '../../../../../../src/shared/domain/models/target-profile-management/StageCollectionUpdate.js';
 
 describe('Unit | Domain | Models | target-profile-management/StageCollectionUpdate', function () {
   describe('constructor', function () {
