+++ api/tests/shared/unit/domain/models/target-profile-management/TargetProfile_test.js
@@ -1,5 +1,5 @@
-import { expect } from '../../../../test-helper.js';
-import { TargetProfile } from '../../../../../lib/domain/models/target-profile-management/TargetProfile.js';
+import { expect } from '../../../../../test-helper.js';
+import { TargetProfile } from '../../../../../../src/shared/domain/models/target-profile-management/TargetProfile.js';
 
 describe('Unit | Domain | Models | TargetProfile', function () {
   describe('#detach', function () {
