+++ api/tests/tooling/domain-builder/factory/target-profile-management/build-stage-collection.js
@@ -1,4 +1,4 @@
-import { StageCollection } from '../../../../../lib/domain/models/target-profile-management/StageCollection.js';
+import { StageCollection } from '../../../../../src/shared/domain/models/target-profile-management/StageCollection.js';
 
 const buildStageCollection = function ({ id, stages = [], maxLevel } = {}) {
   return new StageCollection({ id, stages, maxLevel });
