+++ api/tests/unit/infrastructure/serializers/jsonapi/participant-result-serializer_test.js
@@ -2,7 +2,7 @@ import { domainBuilder, expect } from '../../../../test-helper.js';
 import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/participant-result-serializer.js';
 import { AssessmentResult } from '../../../../../lib/domain/read-models/participant-results/AssessmentResult.js';
 import { KnowledgeElement } from '../../../../../lib/domain/models/index.js';
-import { StageCollection } from '../../../../../lib/domain/models/user-campaign-results/StageCollection.js';
+import { StageCollection } from '../../../../../src/shared/domain/models/user-campaign-results/StageCollection.js';
 
 describe('Unit | Serializer | JSON API | participant-result-serializer', function () {
   context('#serialize', function () {
